Added quantity to listing

+++ db/migrate/20180411191740_add_quantity_to_checkouts.rb
@@ -1,5 +1,5 @@
 class AddQuantityToCheckouts < ActiveRecord::Migration
   def change
-    add_column :checkouts, :quantity, :integer
+    add_column :checkouts, :quantity, :decimal
   end
 end
